add mockSelector to SearchInput tests
