+++ src/__tests__/SearchInput.test.js
@@ -1,12 +1,21 @@
 import React from 'react';
+import * as reactRedux from 'react-redux';
 import { Provider } from 'react-redux';
 import { createStore } from 'redux';
 import { render, fireEvent, cleanup } from '@testing-library/react';
 import '@testing-library/jest-dom/extend-expect';
 import searchInputReducer from '../reducers/searchInput';
 import SearchInput from '../containers/SearchInput';
+import CoinList from '../containers/CoinList';
 
 afterEach(cleanup);
+beforeEach(() => {
+  useSelectorMock.mockClear();
+  useDispatchMock.mockClear();
+});
+
+const useSelectorMock = jest.spyOn(reactRedux, 'useSelector');
+const useDispatchMock = jest.spyOn(reactRedux, 'useDispatch');
 
 function renderWithRedux(
   component,
@@ -28,6 +37,8 @@ it('renders correctly', () => {
 });
 
 describe('Input value', () => {
+  useSelectorMock.mockReturnValue({ CoinList });
+
   it('can update on change', () => {
     const { queryByPlaceholderText } = renderWithRedux(<SearchInput />);
     const searchInput = queryByPlaceholderText(
